Lock simplecov to work around CodeClimate issue
They depend on internal APIs, see:
https://github.com/codeclimate/ruby-test-reporter/issues/180
https://github.com/codeclimate/ruby-test-reporter/pull/181

This is a second rewrite of their test reporter but they don't plan on
fixing this bug because they are working on the third rewrite...

I'm tired of having to deal with CodeClimate breakages on all my
projects so I'll be moving away from it. If you know a good alternative,
**well-written**, with support for aggregating multiple test suites,
please comment on this commit message.