Allow relative keys in any Ruby object
When using relative keys in controllers and mailers, it usually makes
sense to assume the name of the method (such as "index" or "new") should
be part of the absolute key, since they generally follow the "one method
== one action" pattern and there's a view file associated with this
action.

However, in other classes, like presenters, decorators, components, or
any other plain old Ruby object, that isn't usually the case. For
instance, in components the view file is associated with the whole Ruby
class and not just one method.

So now the calling method is only included as part of the key in classes
where it's determined the key should be relative to the method.