Allow formality to be set in config
DeepL allows formality to be optionally more or less formal for certain
languages
(https://www.deepl.com/docs-api/translate-text/translate-text/).
However, the DeepL gem does not allow this to be configured globally,
only on a per-call basis to the `DeepL.translate` call.

By adding a `deepl_options:` key to the configuration, we allow some of
these keys to be overridden.