Remove default translation backend
Modify the Enum option parser to have an option, `default` which allows indicating if the option parser should fall back to the first valid value.

For translation backends, we do not want to have a default because the user must configure the translation backend.

Modify the translation_backend argument to provide an error message when no backend is specified